Move things to 'FundTransactionHandling' trait
