+++ wallet/src/main/scala/org/bitcoins/wallet/LockedWallet.scala
@@ -2,18 +2,10 @@ package org.bitcoins.wallet
 
 import org.bitcoins.core.api.{ChainQueryApi, NodeApi}
 import org.bitcoins.core.bloom.{BloomFilter, BloomUpdateAll}
-import org.bitcoins.core.config.BitcoinNetwork
 import org.bitcoins.core.crypto._
 import org.bitcoins.core.currency._
-import org.bitcoins.core.protocol.transaction.{
-  Transaction,
-  TransactionOutPoint,
-  TransactionOutput
-}
-import org.bitcoins.core.wallet.builder.BitcoinTxBuilder
-import org.bitcoins.core.wallet.fee.FeeUnit
+import org.bitcoins.core.protocol.transaction.TransactionOutPoint
 import org.bitcoins.core.wallet.utxo.TxoState
-import org.bitcoins.keymanager.bip39.BIP39KeyManager
 import org.bitcoins.wallet.api._
 import org.bitcoins.wallet.config.WalletAppConfig
 import org.bitcoins.wallet.internal._
@@ -26,6 +18,7 @@ abstract class LockedWallet
     with UtxoHandling
     with AddressHandling
     with AccountHandling
+    with FundTransactionHandling
     with TransactionProcessing
     with RescanHandling {
 
@@ -61,106 +54,6 @@ abstract class LockedWallet
 
         filtered.fold(0.sats)(_ + _)
       }
-  }
-
-  override def fundRawTransaction(
-      destinations: Vector[TransactionOutput],
-      feeRate: FeeUnit,
-      fromAccount: AccountDb): Future[Transaction] = {
-    val txBuilderF =
-      fundRawTransactionInternal(destinations = destinations,
-                                 feeRate = feeRate,
-                                 fromAccount = fromAccount,
-                                 keyManagerOpt = None)
-    txBuilderF.flatMap(_.unsignedTx)
-  }
-
-  /** This returns a [[BitcoinTxBuilder]] that can be used to generate a unsigned transaction with [[BitcoinTxBuilder.unsignedTx unsignedTx]]
-    * which can be used with signing, or you can just directly call [[BitcoinTxBuilder.sign sign]] to sign the transaction with this instance
-    * of [[BitcoinTxBuilder]]
-    *
-    * If you pass in a [[KeyManager]], the [[org.bitcoins.core.wallet.utxo.UTXOSpendingInfo.signers signers]]
-    * will be populated with really signers that can be used to produce valid [[org.bitcoins.core.crypto.ECDigitalSignature signatures]]
-    *
-    * If you do not pass in a key manager, the transaction will contain [[org.bitcoins.core.protocol.script.EmptyScriptSignature EmptyScriptSignature]]
-    * */
-  private[wallet] def fundRawTransactionInternal(
-      destinations: Vector[TransactionOutput],
-      feeRate: FeeUnit,
-      fromAccount: AccountDb,
-      keyManagerOpt: Option[BIP39KeyManager]): Future[BitcoinTxBuilder] = {
-    val utxosF = listUtxos()
-    val changeAddrF = getNewChangeAddress(fromAccount)
-    val selectedUtxosF = for {
-      walletUtxos <- utxosF
-      change <- changeAddrF
-      //currently just grab the biggest utxos
-      selectedUtxos = CoinSelector
-        .accumulateLargest(walletUtxos, destinations, feeRate)
-    } yield selectedUtxos
-
-    val addrInfosWithUtxoF: Future[Vector[(SpendingInfoDb, AddressInfo)]] =
-      for {
-        selectedUtxos <- selectedUtxosF
-        addrInfoOptF = selectedUtxos.map { utxo =>
-          val addrInfoOptF = getAddressInfo(utxo)
-          //.get should be safe here because of foreign key at the database level
-          addrInfoOptF.map(addrInfoOpt => (utxo, addrInfoOpt.get))
-        }
-        vec <- Future.sequence(addrInfoOptF)
-      } yield vec
-
-    val txBuilderF = for {
-      addrInfosWithUtxo <- addrInfosWithUtxoF
-      change <- changeAddrF
-      utxoSpendingInfos = {
-        addrInfosWithUtxo.map {
-          case (utxo, addrInfo) =>
-            keyManagerOpt match {
-              case Some(km) =>
-                utxo.toUTXOSpendingInfo(account = fromAccount,
-                                        keyManager = km,
-                                        networkParameters = networkParameters)
-              case None =>
-                utxo.toUTXOSpendingInfo(account = fromAccount,
-                                        sign = Sign.dummySign(addrInfo.pubkey),
-                                        networkParameters = networkParameters)
-            }
-
-        }
-      }
-      txBuilder <- {
-
-        logger.info({
-          val utxosStr = utxoSpendingInfos
-            .map { utxo =>
-              import utxo.outPoint
-              s"${outPoint.txId.hex}:${outPoint.vout.toInt}"
-            }
-            .mkString(", ")
-          s"Spending UTXOs: $utxosStr"
-        })
-
-        utxoSpendingInfos.zipWithIndex.foreach {
-          case (utxo, index) =>
-            logger.info(s"UTXO $index details: ${utxo.output}")
-        }
-
-        networkParameters match {
-          case b: BitcoinNetwork =>
-            BitcoinTxBuilder(destinations = destinations,
-                             utxos = utxoSpendingInfos,
-                             feeRate = feeRate,
-                             changeSPK = change.scriptPubKey,
-                             network = b)
-        }
-
-      }
-    } yield {
-      txBuilder
-    }
-
-    txBuilderF
   }
 
   override def getConfirmedBalance(): Future[CurrencyUnit] = {
